Done the Quiz system, needs to be shined and perfumed in the best OOP standards
