+++ DungeonsOfOOP/DungeonsOfOOP/MainGame.cs
@@ -21,26 +21,26 @@ namespace DungeonsOfOOP
 
         static void Main(string[] args)
         {
-            //Opening here to test with random
-            //Also included in Quiz.cs, to be cleaned up but first advicing with teammembers && having a nice evening with friends.
-            var questionsDoc = new XmlDocument();
-            questionsDoc.Load("./../../DataStorage/Questions.xml");
-            XmlNodeList questions = questionsDoc.GetElementsByTagName("question");
-            //questions should be one and only throughout the game to avoid repeating questions. 
+            ////Opening here to test with random
+            ////Also included in Quiz.cs, to be cleaned up but first advicing with teammembers && having a nice evening with friends.
+            //var questionsDoc = new XmlDocument();
+            //questionsDoc.Load("./../../DataStorage/Questions.xml");
+            //XmlNodeList questions = questionsDoc.GetElementsByTagName("question");
+            ////questions should be one and only throughout the game to avoid repeating questions. 
             var random = new Random();
 
             while (true)
             {
-                var questionIndex = random.Next(questions.Count - 1);
-                if (Quiz.AskQuestion(questionIndex, questions))
+                var questionIndex = random.Next(Questionnaire.questions.Count - 1);
+                if (Quiz.AskQuestion(questionIndex, Questionnaire.questions))
                 {
                     Console.WriteLine("True");
-                    questions[questionIndex].ParentNode.RemoveChild(questions[questionIndex]);
+                    Questionnaire.questions[questionIndex].ParentNode.RemoveChild(Questionnaire.questions[questionIndex]);
                 }
                 else
                 {
                     Console.WriteLine("False");
-                    questions[questionIndex].ParentNode.RemoveChild(questions[questionIndex]);
+                    Questionnaire.questions[questionIndex].ParentNode.RemoveChild(Questionnaire.questions[questionIndex]);
                 }
             }
             //Draw images
